feat(terra-draw-maplibre-gl-adapter): preserve map draggability
`setDraggability` was preserving the state of map dragPan and dragRotate
at instantiation time of the adapter. However, if you change map
draggability _outside_ of terradraw after the adapter is instantiated,
then it would not respect that updated value.

This change makes the assumption that terradraw will always attempt to
disable map interactions (in order to handle dragging drawn features),
and then re-enable it later. We use that as a way to determine when to
snapshot/restore the external map drag values.